more web refactoring
