+++ crates/domino-web/static/index.html
@@ -11,6 +11,7 @@
       <p>Please supply a zip of your project! If you don't have any by hand, you can use <a href="hello-world.zip">hello-world</a> or <a href="4WHS-full.zip">4WHS-full</a> to play!</p>
       <input type="file" id="file_upload"></input>
       <input type="button" value="CVC5 Test" id="button_cvctest"></input>
+      <input type="button" value="Domino Test" id="button_dominotest"></input>
       <input type="button" value="Proof Steps" id="button_proof_steps"></input>
       <input type="button" value="Prove" id="button_prove"></input>
